NoC number of nodes and hSize are now of int type (instead on uint32_t)

+++ src/noc/topology/noc-mesh-2d.h
@@ -61,7 +61,7 @@ namespace ns3
     /**
      * how many nodes the 2D mesh will have on one horizontal line
      */
-    uint8_t m_hSize;
+    int m_hSize;
 
   };
